--- file.js
@@ -1,17 +1,0 @@
-// const fs = require('fs');
-
-// //-----FUNCION PARA VER ARCHIVOS MD EN FILE O FOLDER
-// function filesFromPath(routes) {
-//     console.log('hola mundo')
-//   let absolutePath = path.resolve(routes)
-// //   let statsObj = fs.statSync(absolutePath)
-//     // if(statsObj.isFile()) {
-//     //     let extension = path.extname(absolutePath)
-//     //     if (extension == '.md') {
-//     //         return [absolutePath]
-//     //     }
-//     // }
-// }
-// module.exports = {
-//     filesFromPath
-// }
