+++ package.json
@@ -7,5 +7,8 @@
     "file-system": "^2.2.2",
     "node-fetch": "^3.1.0",
     "yargs": "^17.3.1"
+  },
+  "scripts": {
+    "test": "node --experimental-vm-modules ./node_modules/jest/bin/jest.js --coverage"
   }
 }
