+++ prueba.md
@@ -1,2 +1,2 @@
 [Prueba error](https://www.pixar.com/error404) hola error
-[Prueba error](https://www.google.com/) hola google
+[Prueba google](https://www.google.com/) hola google
